+++ src/main/java/com/tnt/ecommeracemarketplace/configuration/JPAConfiguration.java
@@ -1,6 +1,6 @@
 package com.tnt.ecommeracemarketplace.configuration;
 
-import com.tnt.ecommeracemarketplace.repository.ProductRepository;
+import com.tnt.ecommeracemarketplace.products.ProductRepository;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
