+++ src/main/java/com/tnt/ecommeracemarketplace/productEs/ProductEsController.java
@@ -1,8 +1,7 @@
-package com.tnt.ecommeracemarketplace.controller;
+package com.tnt.ecommeracemarketplace.productEs;
 
 import com.tnt.ecommeracemarketplace.dto.PageDto;
 import com.tnt.ecommeracemarketplace.dto.ProductListResponseDto;
-import com.tnt.ecommeracemarketplace.service.ProductEsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
